pyproject.toml
[project]
name = "buse"
version = "0.3.1"
description = "buse: Stateless CLI for browser-use"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "rinvii" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Testing",
]
keywords = ["browser", "automation", "cli", "agents", "browser-use"]
dependencies = [
    "anthropic>=0.75.0",
    "browser-use>=0.11.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "rich>=14.2.0",
    "toon-format>=0.9.0b1",
    "typer>=0.21.1",
    "psutil>=7.2.1",
    "Pillow>=10.0.0",
    "mcp>=1.25.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
buse = "buse.main:app"

[project.urls]
Homepage = "https://github.com/rinvii/buse"
Repository = "https://github.com/rinvii/buse"
Issues = "https://github.com/rinvii/buse/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
